moved interfaces and types to respective folders and imported them wherever required (#126)
