+++ app/types/GlobalState.ts
@@ -1,0 +1,1 @@
+export type updateUserKey = 'firstname' | 'lastname' | 'bio' | 'username' | 'timezone';
